server/questionnaire/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
openpyxl==3.1.2
werkzeug==2.3.7
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
xlrd==2.0.1
OpenAI==1.55.1
httpx==0.25.2
python-dotenv